Moved entire FlutterBuild into flutter module
