+++ src/auto_flutter_dig/module/firebase/task/upload.py
@@ -4,14 +4,10 @@ from ....core.config import Config
 from ....core.os import OS
 from ....core.utils import _Dict, _If
 from ....model.argument.option import LongOptionWithValue
+from ....module.flutter.task.build.stub import FlutterBuildStub
 from ....task.base.process import *
-from ....task.flutter.build.stub import FlutterBuildStub
 from ..identity import FirebaseTaskIdentity
-from ..model._const import (
-    FIREBASE_CONFIG_KEY_PATH,
-    FIREBASE_DISABLE_INTERACTIVE_MODE,
-    FIREBASE_ENV,
-)
+from ..model._const import FIREBASE_CONFIG_KEY_PATH, FIREBASE_DISABLE_INTERACTIVE_MODE, FIREBASE_ENV
 from .setup.check import FirebaseCheck
 from .validate import FirebaseBuildValidate
